Bug 91 - eliminating redundant SStructures for word tokens Feature 92 - adding cite annotation on word tokens - this can contain the cts urn of the citation in which the word is found
Also, removing the transform from beta to unicode from the code. it wasn't working as coded because it requires a Saxon parser, and although I could have fixed that by specifying the parser class, in the end it seems better to keep that out of the importer code because it's not directly related to the process of mapping the perseus format to the salt metamodel.

git-svn-id: https://korpling.german.hu-berlin.de/svn/saltnpepper/PepperModules/PerseusModules/trunk@7211 47085195-c50e-4317-8f43-7cb808db0f00